+++ platform/windows/cs/Adguard.Dns/Adguard.Dns.TestApp/Program.cs
@@ -55,6 +55,9 @@ namespace Adguard.Dns.TestApp
                     m_CoreProcess.Start();
                 }
 
+                dnsProxySettings.HandleDNSSuffixes = true;
+                dnsProxySettings.OptimisticCache = true;
+
                 m_DnsApi.StartDnsFiltering(new DnsApiConfiguration
                 {
                     IsEnabled = true,
@@ -130,6 +133,7 @@ namespace Adguard.Dns.TestApp
                     CreateUpstreamOptions()
                 },
                 Fallbacks = new List<UpstreamOptions>(),
+                UserDNSSuffixes = new List<string>(),
                 Dns64 = new Dns64Settings
                 {
                     Upstreams = new List<UpstreamOptions>(),