+++ platform/windows/cs/Adguard.Dns/Adguard.Dns.Tests/Helpers/ConfigurationHelper.cs
@@ -33,6 +33,7 @@ namespace Adguard.Dns.Tests.Helpers
                 {
                     CreateUpstreamOptions()
                 },
+                UserDNSSuffixes = new List<string>(),
                 Dns64 = new Dns64Settings
                 {
                     Upstreams = new List<UpstreamOptions>