+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## V1.5
+* [Feature] Add ability to redirect all requests with DNS suffix to the fallbacks<p>
+    see `ag::dnsproxy_settings::handle_dns_suffixes`
 * [Feature] DNS-over-QUIC default port changed. New port is 8853.
     Now an address like `quic://dns.adguard.com` is transformed into `quic://dns.adguard.com:8853`
     So to force the use of the old port `784` specify it strictly - `quic://dns.adguard.com:784`.